Re-organize imports according to PEP 8
Since I expect most users to be importing `*` from this module,
we should be using absolute imports for the few standard library functions
we need.

The alternative to this would be putting the os/mimetypes imports within
the `is_image` function, as this would prevent the `from vsutil import *`
user from having these modules' namespaces reserved, but this is not the
proper location for imports according to PEP 8.