Create a _metadata.py file to re-use version number and author
This makes it so the version number only needs to be bumped in one
location (and it's named clearly) instead of 3.

Can be modified in the future if we want to add more fields,
like a date. There's also the possibility to auto-update that file with
git tags but might be a bit overkill for now.